requirements.txt
Django~=3.2
django-import-export